+++ src/layers/custom-tile-3d-layer/mesh-layer/mesh-layer-vertex.glsl.ts
