+++ src/layers/index.ts
@@ -1,1 +1,2 @@
 export { default as BoundingVolumeLayer } from "./bounding-volume-layer/bounding-volume-layer";
+export { default as CustomTile3DLayer } from "./custom-tile-3d-layer/custom-tile-3d-layer";
